se muestra morosidad
