--- frontend/src/pages/Vivienda/vivienda.module.css
@@ -1,134 +1,0 @@
-.container {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 20px;
-}
-
-.title {
-  color: #333;
-  text-align: center;
-  margin-bottom: 30px;
-}
-
-.section {
-  background: white;
-  border-radius: 8px;
-  padding: 20px;
-  margin-bottom: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.section h3 {
-  color: #2c3e50;
-  margin-bottom: 20px;
-  padding-bottom: 10px;
-  border-bottom: 2px solid #eee;
-}
-
-.viviendaInfo {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 15px;
-}
-
-.viviendaInfo p {
-  margin: 8px 0;
-}
-
-.habitantesList {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 20px;
-}
-
-.habitanteCard {
-  background: #f8f9fa;
-  border-radius: 8px;
-  padding: 15px;
-  border: 1px solid #dee2e6;
-}
-
-.habitanteCard h4 {
-  color: #2c3e50;
-  margin-bottom: 15px;
-  padding-bottom: 8px;
-  border-bottom: 1px solid #dee2e6;
-}
-
-.habitanteInfo p {
-  margin: 5px 0;
-}
-
-.ingresosSection {
-  margin-top: 15px;
-  padding-top: 15px;
-  border-top: 1px solid #dee2e6;
-}
-
-.ingresoItem {
-  background: white;
-  padding: 10px;
-  margin: 10px 0;
-  border-radius: 4px;
-  border: 1px solid #eee;
-}
-
-.totalIngresos {
-  margin-top: 15px;
-  padding-top: 10px;
-  border-top: 1px dashed #dee2e6;
-  font-size: 1.1em;
-  color: #28a745;
-}
-
-.loteInfo {
-  margin-top: 15px;
-  padding-top: 15px;
-  border-top: 1px solid #dee2e6;
-}
-
-.totalVivienda {
-  margin-top: 30px;
-  padding: 20px;
-  background: #28a745;
-  color: white;
-  border-radius: 8px;
-  text-align: center;
-}
-
-.loading {
-  text-align: center;
-  padding: 50px;
-  font-size: 1.2em;
-  color: #666;
-}
-
-.error {
-  text-align: center;
-  padding: 50px;
-  color: #dc3545;
-}
-
-.morosidadSection {
-  margin-top: 15px;
-  padding-top: 15px;
-  border-top: 1px solid #DEE2E6;
-}
-
-.morosidadSection_italic{
-  font-style: italic;
-}
-
-@media (max-width: 768px) {
-  .container {
-    padding: 10px;
-  }
-
-  .habitantesList {
-    grid-template-columns: 1fr;
-  }
-
-  .viviendaInfo {
-    grid-template-columns: 1fr;
-  }
-}
